requirements.txt
yfinance
selenium
webdriver-manager
beautifulsoup4
rich
google-generativeai
python-dotenv
requests